requirements.txt
mypy
aiohttp
elasticsearch
torch
transformers
Markdown
pyyaml
attr
